requirements.txt
httpx>=0.24.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
